// Web/src/app/models/DiseaseVM.ts
export class DiseaseVM{
    ID_disease: number;
    name: string;
    description: string;
    type: string;
    severity: string;
    spread: string;
    symptoms: string;
    prevention: string;
    remedy: string;
    company: number;
}

export class SearchDiseaseVM{
    TotalRows: number;
    ID_disease: number;
    name: string;
    description: string;
    type: string;
    severity: string;
    spread: string;
    symptoms: string;
    prevention: string;
    remedy: string;
    company: number;
}